% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{calculate cell*cell cosine similarity}
\usage{
cosine_similarity(matrix)
}
\arguments{
\item{matrix}{input matrix (rows = cells, column = genes)}
}
\value{
cells*cells matrix
}
\description{
calculate cell*cell cosine similarity
}
