% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{countFilter}
\alias{countFilter}
\title{Count filters and plots}
\usage{
countFilter(
  counts,
  minGeneCount = 0,
  maxGeneCount = Inf,
  minCellCount = 100,
  minCellsForGenes = 0,
  minGenesForCells = 0
)
}
\arguments{
\item{counts}{long format count matrix (with columns: cell, gene, count)}

\item{minGeneCount}{min Counts for a gene to be kept}

\item{maxGeneCount}{max Counts for a gene to be kept (remove outliers)}

\item{minCellCount}{min Counts for a Cell to be kept}

\item{minCellsForGenes}{to filter genes, min number of cells with nonzero counts}

\item{minGenesForCells}{to filter cells, min number of genes with nonzero counts}
}
\value{
a list, with filtered geneCounts in tibble (cell,gene, count), and wide matrix (genes*cells)
}
\description{
Count filters and plots
}
\examples{

}
