% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_and_importing.R
\name{countsInRegions_perCell}
\alias{countsInRegions_perCell}
\title{get counts in a given region (GRanges) per cell from BAM files}
\usage{
countsInRegions_perCell(regions, barcodelist, bamFiles)
}
\arguments{
\item{regions}{GRanges of regions to count}

\item{barcodelist}{vector, barcode list}

\item{bamFiles}{bam files path}
}
\value{
counts per region per cell
}
\description{
get counts in a given region (GRanges) per cell from BAM files
}
