% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{countToEffCounts}
\alias{countToEffCounts}
\title{Convert counts to Effective counts}
\usage{
countToEffCounts(counts, len, effLen)
}
\arguments{
\item{counts}{vector of counts per gene}

\item{len}{actual length per gene}

\item{effLen}{effective length per gene}
}
\value{
vector with effective counts
}
\description{
Convert counts to Effective counts
}
