% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{countToFpkm}
\alias{countToFpkm}
\title{Convert counts to FPKM}
\usage{
countToFpkm(counts, effLen)
}
\arguments{
\item{counts}{Count matrix of genes*samples}

\item{effLen}{vector of effective gene lengths per gene}
}
\value{
matrix of RPKM normalized counts
}
\description{
Convert counts to FPKM
}
