% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{countToTpm}
\alias{countToTpm}
\title{Convert counts to TPM (or TP<scale_factor>)}
\usage{
countToTpm(counts, effLen, scale_factor = 1e+06)
}
\arguments{
\item{counts}{Count matrix of genes*samples}

\item{effLen}{vector of effective gene lengths per gene}
}
\value{
matrix of TPM normalized counts
}
\description{
Convert counts to TPM (or TP<scale_factor>)
}
