% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transformations.R
\name{cpm_bygroup}
\alias{cpm_bygroup}
\title{Obtain average CPM per group from a SingleCellExperiment object}
\usage{
cpm_bygroup(sce, group_var, assays, aggtype = "mean")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{group_var}{variable in colData(sce) that can be used to group cells}

\item{assays}{slot from assayNames(sce) which we want to transform}

\item{aggtype}{type of aggregation (sum/mean)}
}
\value{
dataframe with genes/regions in rows and groups in columns
}
\description{
Obtain average CPM per group from a SingleCellExperiment object
}
