% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transformations.R
\name{deseq_norm}
\alias{deseq_norm}
\title{Normalize a single-cell count matrix based on DEseq sizefactor}
\usage{
deseq_norm(m)
}
\arguments{
\item{m}{matrix (cells in cols, genes in rows)}
}
\value{
matrix (normalized)
}
\description{
Normalize a single-cell count matrix based on DEseq sizefactor
}
