% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF-IDF_functions.R
\name{do_pca}
\alias{do_pca}
\title{Run fast PCA (irlba) on a feature*sample matrix}
\usage{
do_pca(mat, dims = 50)
}
\arguments{
\item{mat}{feature*sample matrix (eg. rows= genes/regions, cols=cells), counts must already be normalized and scaled}

\item{dims}{how many dims to calculate?}
}
\value{
list of feature*PC and cell\*PC output matrices
}
\description{
Run fast PCA (irlba) on a feature*sample matrix
}
