% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAnno.R
\name{dotAnno}
\alias{dotAnno}
\title{Generate DotPlot with informative details to assist with cluster annotation}
\usage{
dotAnno(obj, clusters = "all", type = "all")
}
\arguments{
\item{obj}{Seurat scRNA-seq object (default assay must be RNA or SCT, already normalized)}

\item{clusters}{clusters of interest to provide dotplot (default is all)}

\item{type}{A string specifiying the type of annotation}
}
\value{
A ggplot object of dotplot with annotations
}
\description{
Generate DotPlot with informative details to assist with cluster annotation
}
