% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAfraction.R
\name{filterReadMotifs}
\alias{filterReadMotifs}
\title{filter Reads based on nucleotide content at 5-prime}
\usage{
filterReadMotifs(
  bamdf,
  pos_strand_expected,
  neg_strand_expected,
  plotFile = NA,
  bsgenome
)
}
\arguments{
\item{bamdf}{bam file as data frame}

\item{pos_strand_expected}{expected Nucleoties in read and genome for alignments in + strand c("A", "AT")}

\item{neg_strand_expected}{expected Nucleoties in read and genome for alignments in - strand c("A", "AT")}

\item{plotFile}{file to save the plot of nucleotide content}

\item{bsgenome}{a BSgenome object}
}
\description{
filter Reads based on nucleotide content at 5-prime
}
