% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIC_read_filtering.R
\name{get5pNucCounts}
\alias{get5pNucCounts}
\title{Get counts for given motif at 5'-cut-site of MNAse}
\usage{
get5pNucCounts(
  bamfile,
  plus_strand_expected = c("A", "TA"),
  minus_strand_expected = c("T", "TA"),
  plot = NA,
  bsGenome = NA
)
}
\arguments{
\item{bamfile}{bam file object (using Rsamtools::bamFile)}

\item{plus_strand_expected}{vector. expected motif at forward-oriented reads and their genomic location}

\item{minus_strand_expected}{vector. expected motif at forward-oriented reads and their genomic location}

\item{plot}{fileName to plot}

\item{bsGenome}{bsGenome object}
}
\value{
numeric with no. of reads passing the filter
}
\description{
Get counts for given motif at 5'-cut-site of MNAse
}
