% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_colors}
\alias{get_colors}
\title{Return hex codes for distinct colors mapped to input names}
\usage{
get_colors(names, color_space = NA)
}
\arguments{
\item{names}{vector of names (each unique name is mapped to a hex code)}
}
\value{
clist vector of colors which can be used in ggplot2::scale_color_manual
}
\description{
Return hex codes for distinct colors mapped to input names
}
