% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAfraction.R
\name{getDistanceBetweenCuts}
\alias{getDistanceBetweenCuts}
\title{Get distance between R1-cuts on opposite strand per cell}
\usage{
getDistanceBetweenCuts(
  bam,
  barcodes,
  subsetRegions = NA,
  nCores = 2,
  linePlot = "fragSizes_from_scCuts.png",
  heatmap = "fragSizes_from_scCuts_heatmap.png"
)
}
\arguments{
\item{bam}{BAM file}

\item{barcodes}{barcodes (chr vector)}

\item{subsetRegions}{GRanges to subset to (if needed)}

\item{nCores}{Int, No. of cores}

\item{linePlot}{chr, name of line plot}

\item{heatmap}{chr, name of heatmap}
}
\value{
data.frame with barcode and distance between cuts
}
\description{
Get distance between R1-cuts on opposite strand per cell
}
