% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transformations.R
\name{getGeneActivity}
\alias{getGeneActivity}
\title{Collate binCounts at gene level based on overlap}
\usage{
getGeneActivity(mat, genes_gr, subset = NULL, overlapGap = 10000)
}
\arguments{
\item{mat}{Count matrix (bin-level) with regions as rownames (format: chr_start_end)}

\item{genes_gr}{Genes as GRanges object, column "gene_id" should indicate gene names/IDs}

\item{subset}{List of genes to subset from genes_gr for overlap}

\item{overlapGap}{Max distance between gene and bin for overlap}
}
\value{
Sparse matrix (dgcMatrix) with collates counts (genes=rows, cells=columns)
}
\description{
Collate binCounts at gene level based on overlap
}
