% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transformations.R
\name{getHVGs}
\alias{getHVGs}
\title{Get top N highly variable features based on (analytic) pearson residuals}
\usage{
getHVGs(mat, top_n = 3000, plot = FALSE)
}
\arguments{
\item{mat}{sparse matrix (dgCMatrix) of raw counts}

\item{top_n}{numeric, no. of high var features to fetch}

\item{plot}{bool, whether to make mean-variance plot with highly variable features}
}
\value{
vector of rownames(mat)
}
\description{
Get top N highly variable features based on (analytic) pearson residuals
}
