% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{getTopicCorrelations}
\alias{getTopicCorrelations}
\title{Extract topic correlations from a CTM model output}
\usage{
getTopicCorrelations(ctm_gamma, ntopics, plot = TRUE)
}
\arguments{
\item{ctm_gamma}{gamma matrix output of CTM}

\item{ntopics}{number of topics}

\item{plot}{logical. Whether to plot topic*topic correlation across documents.}
}
\value{
data frame of correlations (+plot object if plot= TRUE)
}
\description{
Extract topic correlations from a CTM model output
}
