% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{ldaWrapper}
\alias{ldaWrapper}
\title{Wrapper for LDA, UMAP and clustering}
\usage{
ldaWrapper(geneCounts, binarize = FALSE, nTopic = 20, nK = 30)
}
\arguments{
\item{geneCounts}{geneCounts in long format (gene, cell, counts)}

\item{nTopic}{No. of topics for LDA}

\item{nK}{No. of nearest neighbours for louvain and UMAP}
}
\value{
list with LDA and umap output
}
\description{
Wrapper for LDA, UMAP and clustering
}
