% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_colorList}
\alias{load_colorList}
\title{Return n distinct color as hex code}
\usage{
load_colorList(n)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Return n distinct color as hex code
}
\examples{


}
