% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadExample}
\alias{loadExample}
\title{Load example Seurat datasets for testing}
\usage{
loadExample(ex, norm = FALSE)
}
\arguments{
\item{ex}{Seurat example dataset to load, options include: "3k"}

\item{norm}{A boolean specifying whether to normalize RNA assay and cluster (default FALSE)}
}
\value{
Seurat object of example dataset
}
\description{
Load example Seurat datasets for testing
}
\examples{
 
obj = loadExample("3k")

}
