% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{loadLDAparams}
\alias{loadLDAparams}
\title{Load params for LDA or CTM}
\usage{
loadLDAparams(out)
}
\arguments{
\item{out}{which param to load ('LDA' or 'CTM')}
}
\value{
list of params
}
\description{
Load params for LDA or CTM
}
\examples{

}
