% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadPackages}
\alias{loadPackages}
\title{Load commonly used packages}
\usage{
loadPackages(category = "general")
}
\arguments{
\item{category}{choose from "general", "topicmodels", "stats",
"granges", "rna" or "all"}
}
\value{
Nothing. Just loads packages
}
\description{
Load commonly used packages
}
