% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{look}
\alias{look}
\title{Quick look into a large matrix/df}
\usage{
look(m)
}
\arguments{
\item{m}{matrix/ dataframe}
}
\value{
matrix with 10 rows and cols
}
\description{
Quick look into a large matrix/df
}
