% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_and_importing.R
\name{loom_to_sce}
\alias{loom_to_sce}
\title{Import a loom file as SingleCellExperiment using loomR API}
\usage{
loom_to_sce(
  loomfile,
  obs_names = "obs_names",
  var_names = "var_names",
  reducedDimNames = c("X_pca", "X_umap")
)
}
\arguments{
\item{loomfile}{Path to loom file}

\item{obs_names}{Attribute that stores observation IDs (eg. cell IDs) in loom file}

\item{var_names}{Attribute that stores variable IDs (eg. gene IDs) in loom file}

\item{reducedDimNames}{PCA/UMAP attribute names (from adata.obsm.keys())}
}
\value{
SingleCellExperiment object
}
\description{
This function is only needed in case the loom file contains
  "obsm" and "varm" entries, which causes loomExperiment to error. Otherwise
  simply use LoomExperiment::import(). In this case the "obsm" and "varm"
  entries are added to rowData and colData.
}
