% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTx.R
\name{LR_fit}
\alias{LR_fit}
\title{Gene-wise likelihood ratio test}
\usage{
LR_fit(counts, genes, cell_group, filter_threshold)
}
\arguments{
\item{counts}{tibble of counts (gene, cell, counts)}

\item{genes}{vector with gene names (for testing)}

\item{cell_group}{vector with cell groups (cluster labels)}

\item{filter_threshold}{lower count threshold}
}
\value{
data frame with pvalue per gene
}
\description{
Gene-wise likelihood ratio test
}
\examples{


}
