% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_workflows.R
\name{lsaWorkflowVB}
\alias{lsaWorkflowVB}
\title{vivekbhr's workflow for LSA and annotation}
\usage{
lsaWorkflowVB(binCounts_filt, genes_gr, qc_df = NA)
}
\arguments{
\item{binCounts_filt}{Filtered counts in genomic bins (rows = regions, columns = cells)}

\item{genes_gr}{GenomicRanges with gene names and their locations.}

\item{qc_df}{(optional) DataFrame with per-cell quality control (or other) information}
}
\value{
Annotated LSA object (as list)
}
\description{
vivekbhr's workflow for LSA and annotation
}
\examples{

}
