% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{makeKNNgraph}
\alias{makeKNNgraph}
\title{Convert a cellxPC or a cellxTopic matrix to a KNN matrix}
\usage{
makeKNNgraph(mat, nk, replace_dist = FALSE, method = "RANN")
}
\arguments{
\item{mat}{Input matrix after dim reduction (cell x PC or cell x Topic)}

\item{nk}{no. of K}

\item{replace_dist}{add the distance back to the KNN graph, if TRUE, the distance
to the K-nearest neighbours are retained in the output graph (weighted)
otherwise the distances are removed, returning unweighted graph}

\item{method}{KNN building method (RANN or fnn)}
}
\value{
igraph objet
}
\description{
Convert a cellxPC or a cellxTopic matrix to a KNN matrix
}
\examples{


}
