% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_workflows.R
\name{monocleDNAworkflow}
\alias{monocleDNAworkflow}
\title{Monocle ATAC/DNA workflow}
\usage{
monocleDNAworkflow(counts, nDim = 20)
}
\arguments{
\item{counts}{matrix: either wide (regions*cells) or tidy (regions, cell, count) format}

\item{isTidy}{bool: are the counts in tidy format?}
}
\value{
cellDataSet object
}
\description{
Monocle ATAC/DNA workflow
}
\examples{


}
