% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{plotGamma}
\alias{plotGamma}
\title{Plot the topic to document/cell (gamma) distribution from LDA/CTM output}
\usage{
plotGamma(tidymodel)
}
\arguments{
\item{tidymodel}{tidy output of LDA/CTM}
}
\value{
ggplot object
}
\description{
Plot the topic to document/cell (gamma) distribution from LDA/CTM output
}
\examples{

}
