% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plotGenesPerCluster}
\alias{plotGenesPerCluster}
\title{Plot the counts for gene lists to compare bw clusters}
\usage{
plotGenesPerCluster(geneList, counts, umap, cl_col = "louvain")
}
\arguments{
\item{geneList}{list with names = clusterIDs and values = geneIDs}

\item{counts}{count tibble (gene, cell, count)}

\item{umap}{umap output (rownames=  cells)}

\item{cl_col}{which column contain cluster info in the umap?}
}
\value{
ggplot objects
}
\description{
Plot the counts for gene lists to compare bw clusters
}
\examples{


}
