% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMarker}
\alias{plotMarker}
\title{Plot marker genes on a UMAP}
\usage{
plotMarker(umap, count_df, geneID, title)
}
\arguments{
\item{umap}{a data frame with rownames = cell id, and 2 columns ("UMAP1", "UMAP2")}

\item{count_df}{a long format data frame with 3 columns ("cell", "gene" and "count")}

\item{geneID}{ID of gene to plot in the UMAP}

\item{title}{title of the plot}
}
\value{
ggplot object
}
\description{
Plot marker genes on a UMAP
}
\examples{

}
