% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPlateLayout}
\alias{plotPlateLayout}
\title{Plot numeric information on a 384-well plate layout}
\usage{
plotPlateLayout(barcodeFile, bcInfo, varToPlot = "count", groupInfo = NA)
}
\arguments{
\item{barcodeFile}{File with barcodes arranged in order: 1-384}

\item{bcInfo}{Data frame with row.names=barcodes (matching barcodeFile) and columns with
values to plot (eg. total counts)}

\item{varToPlot}{Which column of bcInfo to plot}

\item{groupInfo}{(optional) list which highlights certain wells. This can be used to provide
well grouping information. For example, if last 4 columns of a plate contain
control samples, or if last 8 corner wells are left empty etc. The list should
contain 3 values row (top to bottom, numeric), col (left to right, numeric), and
group (character/factor).}
}
\value{
Plate plot (ggplot object)
}
\description{
Plot numeric information on a 384-well plate layout
}
\examples{


}
