% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRegionDetection}
\alias{plotRegionDetection}
\title{Plot no. of regions detected as a function of counts}
\usage{
plotRegionDetection(mat)
}
\arguments{
\item{matrix}{A region * cell matrix of counts}
}
\value{
ggplot object
}
\description{
Plot no. of regions detected as a function of counts
}
\examples{


}
