% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plotTopicProbs}
\alias{plotTopicProbs}
\title{PLot topic probability on UMAP}
\usage{
plotTopicProbs(ldaOut, nTopic = 20, outPdf)
}
\arguments{
\item{ldaOut}{output of lda Wrapper}

\item{nTopic}{nTopics to use}

\item{outPdf}{output pdf file}
}
\value{
pdf file with ggplot of topic probs
}
\description{
PLot topic probability on UMAP
}
