% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotVarExplained}
\alias{plotVarExplained}
\title{Plot variance explained for a PC matrix}
\usage{
plotVarExplained(mat)
}
\arguments{
\item{matrix}{N*PC matrix}
}
\value{
ggplot object
}
\description{
Plot variance explained for a PC matrix
}
