% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF-IDF_functions.R
\name{quick_umap_louvain}
\alias{quick_umap_louvain}
\title{Quick UMAP and louvain cluster of a feature*PC data frame}
\usage{
quick_umap_louvain(x, PCs = 2:50, nK = 30, spread = 5, mindist = 0.1)
}
\arguments{
\item{x}{data frame with rownames=features/samples, colnames=PCs}

\item{PCs}{vector of which principal components (columns) to select}

\item{nK}{No. of neighbors for louvain clustering}

\item{spread}{UMAP spread}

\item{mindist}{UMAP min_dist}
}
\value{
data.frame with UMAP1/2 and louvain clusters
}
\description{
Quick UMAP and louvain cluster of a feature*PC data frame
}
