% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{regionDetectionPlot}
\alias{regionDetectionPlot}
\title{Regions detected as a function of counts}
\usage{
regionDetectionPlot(mat)
}
\arguments{
\item{matrix}{A region * cell matrix of counts}
}
\value{
ggplot object
}
\description{
Regions detected as a function of counts
}
\examples{


}
