% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sanitizeGeneID}
\alias{sanitizeGeneID}
\title{Remove version name from ensembl gene IDs}
\usage{
sanitizeGeneID(geneIDs)
}
\arguments{
\item{geneIDs}{char vector: ensembl gene IDs}
}
\value{
char vecter of santized gene IDs
}
\description{
Remove version name from ensembl gene IDs
}
