% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_workflows.R
\name{seuratWorkflow}
\alias{seuratWorkflow}
\title{Seurat workflow}
\usage{
seuratWorkflow(counts)
}
\arguments{
\item{counts}{count matrix (rows = regions, columns = cells)}
}
\value{
seurat object
}
\description{
Seurat workflow
}
