% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF-IDF_functions.R
\name{tfidf}
\alias{tfidf}
\title{Perform TF-IDF transformation of a matrix}
\usage{
tfidf(
  bmat,
  frequencies = TRUE,
  log_scale_tf = TRUE,
  scale_factor = 1e+05,
  threshold_for_idf = 1
)
}
\arguments{
\item{bmat}{sparse matrix of class dgcMatrix (cells = columns, genes/regions = rows)}

\item{frequencies}{logical, divide bmat by colSums (if FALSE simply use bmat for TF matrix)}

\item{log_scale_tf}{logical, take log1p of the term-frequencies?}

\item{scale_factor}{size factor to multiply with (if converting to frequencies)}

\item{threshold_for_idf}{count threshold to convert data to binary. Use for IDF
(IDF is not be used with non-binary data)}
}
\value{
tf-idf (sparse) matrix
}
\description{
Perform TF-IDF transformation of a matrix
}
