% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{tidy_ctm_beta}
\alias{tidy_ctm_beta}
\title{Get the word/region to topic (beta) distribution from a CTM output}
\usage{
tidy_ctm_beta(CTM_object)
}
\arguments{
\item{CTM_object}{output of topicmodels::CTM function}
}
\value{
tibble with beta values
}
\description{
Get the word/region to topic (beta) distribution from a CTM output
}
