% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_CTM_functions.R
\name{tidy_ctm_gamma}
\alias{tidy_ctm_gamma}
\title{Get the topic to document/cell (gamma) distribution from a CTM output}
\usage{
tidy_ctm_gamma(CTM_object)
}
\arguments{
\item{CTM_object}{output of topicmodels::CTM function}
}
\value{
tibble with gamma values
}
\description{
Get the topic to document/cell (gamma) distribution from a CTM output
}
\examples{


}
