% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{topGenesByCluster}
\alias{topGenesByCluster}
\title{Get top-genes per cluster based on T-test}
\usage{
topGenesByCluster(cluster_df, count_matrix, n = NA)
}
\arguments{
\item{cluster_df}{Data frame with 2 columns "cell ID", "cluster no."}

\item{count_matrix}{Wide format count matrix with rownames = genes, colnames = cells}

\item{n}{No. of top genes to extract (NA = return df with all significant genes at padj < 0.05)}
}
\value{
vector of top genes, or (if n = NA) data frame with t-test output
}
\description{
Get top-genes per cluster based on T-test
}
