% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transformations.R
\name{tpt_bygroup}
\alias{tpt_bygroup}
\title{Get transcript per 10k counts per group}
\usage{
tpt_bygroup(sce, group_var, length_var, assays, aggtype = "sum")
}
\arguments{
\item{sce}{SingleCellExperiment object, with cell groups and lengths, counts should be raw}

\item{group_var}{column in colData(sce) to make cell groups}

\item{length_var}{column in rowData(sce) which indicate gene lengths}

\item{assays}{which assays (from assayNames(sce)) to use}

\item{aggtype}{either "sum" (sum by group then take tpt), or "mean" (take tpt per cell then mean by group)}
}
\value{
data.frame (rows=genes, col = groups)
}
\description{
Get transcript per 10k counts per group
}
