% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{complexDimPlot}
\alias{complexDimPlot}
\title{Make a DimPlot (UMAP/PCA) from a SingleCellExperiment Object,
where cluster labels are numeric, and the annotation is separated}
\usage{
complexDimPlot(
  sce,
  dimtoplot = "X_umap",
  vartoplot = "hour",
  point_size = 1.5,
  point_stroke = 0.05,
  numeric_annotation = TRUE,
  separate_legend = FALSE,
  manual_colors = NULL,
  return_objects = FALSE
)
}
\arguments{
\item{dimtoplot}{slot from reducedDimNames(sce) to plot}

\item{vartoplot}{column from colData(sce) to plot (must be categorical)}

\item{point_size}{numeric, size of the points}

\item{point_stroke}{numeric, stroke size of the points}

\item{numeric_annotation}{set it to TRUE to assign numeric IDs to each annotation (this avoids label overlap)}

\item{separate_legend}{if TRUE, legend is plotted in a new plot. By default
legend is stitched to the right of the plot.}

\item{manual_colors}{vector A named vactor of colors, where names = categories present in "vartoplot" column, values are colors}

\item{return_objects}{boolean, whether to return plot objects as list, instead of plotting}

\item{object}{A SingleCellExperiment object (with reducedDim slot),
or a data.frame with with colnames: "Dim1", "Dim2", that are to be plotted.
There must be additional columns present in the input data frame or in colData(sce),
that would be used for coloring/labelling the plot.}
}
\value{
ggplot object
}
\description{
Make a DimPlot (UMAP/PCA) from a SingleCellExperiment Object,
where cluster labels are numeric, and the annotation is separated
}
