% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF-IDF_functions.R
\name{do_pca}
\alias{do_pca}
\title{Run fast PCA (irlba) on TF-IDF matrix}
\usage{
do_pca(mat, dims = 50)
}
\arguments{
\item{mat}{matrix}

\item{dims}{how many dims to calculate?}
}
\value{
PCA output matrix
}
\description{
Run fast PCA (irlba) on TF-IDF matrix
}
