% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prepForPseudobulk}
\alias{prepForPseudobulk}
\title{prepare .txt file to create pseudobulk bigwigs from clustered UMAP (using snakemake workflow)}
\usage{
prepForPseudobulk(umap, prefix, suffix = ".txt")
}
\arguments{
\item{umap}{umap data frame, from standard LSA/LDA wrapper}

\item{prefix}{file name prefix (match it with bam file prefix to run snakemake afterwards)}

\item{suffix}{file name suffix (match it with bam file suffix to run snakemake afterwards)}
}
\value{
None, file is written as prefix_<sampleID>_suffix.txt, which contains cell barcode, and clusterID
}
\description{
prepare .txt file to create pseudobulk bigwigs from clustered UMAP (using snakemake workflow)
}
