% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAnno.R
\name{scAnno}
\alias{scAnno}
\title{Perform SingleR annotation on Seurat scRNA-seq Object}
\usage{
scAnno(obj, type)
}
\arguments{
\item{obj}{Seurat scRNA-seq object (default assay must be RNA, already normalized)}

\item{type}{A string specifying the type of annotation labels to produce (fine or broad labels)}
}
\value{
A vector of Labels to add into Seurat Object
}
\description{
Perform SingleR annotation on Seurat scRNA-seq Object
}
\examples{
cellTypePredictions = scAnno(obj = seuratObj, type = "fine")
seuratObj$cellAnnotation = cellTypePredictions
}
