% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_and_importing.R
\name{totalCounts_perCell}
\alias{totalCounts_perCell}
\title{Get total counts per barcode from BAM files}
\usage{
totalCounts_perCell(barcodelist, bamFiles)
}
\arguments{
\item{barcodelist}{vector with barcodes}

\item{bamFiles}{bamfiles (path)}
}
\value{
data frame with counts per cell
}
\description{
Get total counts per barcode from BAM files
}
